requirements.txt
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
colorama==0.4.4
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.7.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.0.1
monkeylearn==3.5.2
nltk==3.6.2
numpy==1.21.0
oauthlib==3.1.1
pandas==1.2.5
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.1
requests==2.25.1
requests-oauthlib==1.3.0
rfc3986==1.5.0
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sniffio==1.2.0
threadpoolctl==2.1.0
tqdm==4.61.1
tweepy==3.10.0
urllib3==1.26.6
